requirements.txt
streamlit
pandas
numpy
anthropic
python-dotenv
scikit-learn
